fastapi/requirements.txt
fastapi
uvicorn
psycopg2-binary
mlflow
scikit-learn
pandas
numpy
sqlalchemy